services/auth_service/requirements.txt
# Auth Service Requirements
# ========================

# Core Dependencies
fastapi==0.108.0
uvicorn==0.27.1
python-dotenv==1.0.0
pydantic==2.6.1
pydantic[email]>=2.5.0,<2.10.0
pydantic-settings==2.1.0
greenlet==3.0.1

# Authentication
python-jose==3.3.0
python-multipart==0.0.7
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Database & ORM
sqlalchemy==2.0.26
alembic>=1.13.0
asyncpg==0.29.0
aiosqlite==0.19.0  # For SQLite async support in testing

# Networking & Integration
httpx>=0.25.1
requests>=2.28.0

# Testing Dependencies
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
